+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # master
 
+* Fix song_ended event never getting fired
 * Fix laptop names not being able to be updated
 
 ## 0.0.1 / 2012-11-20